Moved QtiVersionException class close to Version class.
